+++ src/main/java/BlockStorage/SSD.java
@@ -18,25 +18,25 @@ public class SSD{
 	private String SSD_LOCATION;
 	private HDFSLayer HDFSLayer;
 	private Utils utils;
-	LinkedHashMap<Long, Boolean> recencyList;
+	LinkedHashMap<Integer, Boolean> recencyList;
 	Lock recencyListLock = new ReentrantLock();
-	CopyOnWriteArrayList<Long> pointersList;
+	CopyOnWriteArrayList<Integer> pointersList;
 
 
-	static Map.Entry<Long, Boolean> elder = null;
+	static Map.Entry<Integer, Boolean> elder = null;
 //	AtomicInteger size = new AtomicInteger(0);
 
-	ConcurrentLinkedQueue<Pair<Long,Integer>> WritetoSSDqueue = new ConcurrentLinkedQueue<>();
-	ConcurrentLinkedQueue<Long> WritetoHDFSqueue = new ConcurrentLinkedQueue<>();
+	ConcurrentLinkedQueue<Pair<Integer,Integer>> WritetoSSDqueue = new ConcurrentLinkedQueue<>();
+	ConcurrentLinkedQueue<Integer> WritetoHDFSqueue = new ConcurrentLinkedQueue<>();
 
 	SSD(HDFSLayer HDFSLayer, Utils utils){
 		this.HDFSLayer = HDFSLayer;
 		this.utils = utils;
 		setSSD_LOCATION();
-		this.recencyList = new LinkedHashMap<Long, Boolean>(utils.SSD_SIZE, 0.75F, false) {
+		this.recencyList = new LinkedHashMap<Integer, Boolean>(utils.SSD_SIZE, 0.75F, false) {
 			private static final long serialVersionUID = 1L;
 			@Override
-			protected boolean removeEldestEntry(Map.Entry<Long, Boolean> eldest) {
+			protected boolean removeEldestEntry(Map.Entry<Integer, Boolean> eldest) {
 				elder =  eldest;
 				return size() > utils.SSD_SIZE;
 			}
@@ -49,13 +49,13 @@ public class SSD{
 	}
 
 
-	void writeSSDPage(long pageNumber, int pointer) {
+	void writeSSDPage(int pageNumber, int pointer) {
 		WritetoSSDqueue.add(new Pair<>(pageNumber, pointer));
 		if(utils.SHOW_LOG)
 			System.out.println("page " + pageNumber + " added to WritetoSSDqueue");
 	}
 
-	page readSSDPage(long pageNumber){
+	page readSSDPage(int pageNumber){
 		File file = new File(SSD_LOCATION + "/" + pageNumber);
 		byte[] pageData = new byte[utils.PAGE_SIZE];
 		try
@@ -76,7 +76,7 @@ public class SSD{
 		return new page(pageNumber, pageData);
 	}
 
-	page readPage(long pageNumber) {
+	page readPage(int pageNumber) {
 		recencyListLock.lock();
 		recencyList.remove(pageNumber);
 		recencyList.put(pageNumber, true);
@@ -119,7 +119,7 @@ public class SSD{
 //		file.delete(); //remove the file from actual SSD
 //	}
 
-	void writePage(long pageNumber,int pointer, blockServer server){
+	void writePage(int pageNumber,int pointer, blockServer server){
 
 		while (pointersList.size() >= utils.SSD_SIZE){
 //			System.out.println("waiting for SSD to have space");
